refactor(ping_pong/game): remove assets.hpp fix(ping_pong/game/README.md): fix run command

+++ ping_pong/game/src/gameobj/ball.cpp
@@ -1,6 +1,6 @@
 #include "gameobj/ball.hpp"
-#include "assets.hpp"
 #include <vector>
+#include "screen.hpp"
 
 Ball::Ball(float _speedX, float _speedY, int _width, int _height)
     : width(_width), height(_height), defaultSpeedX(_speedX),
@@ -15,8 +15,8 @@ void Ball::tick(struct scrsize ws, size_t frameTimeMs) {
 }
 
 void Ball::reset(struct scrsize ws) {
-    this->posX = ws.width / 2;
-    this->posY = ws.height / 2;
+    this->posX = static_cast<float>(ws.width) / 2;
+    this->posY = static_cast<float>(ws.height) / 2;
 
     this->speedX = this->defaultSpeedX;
     this->speedY = this->defaultSpeedY;
